requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.26.0
aws-lambda-powertools>=2.25.0
pydantic>=2.0.0